Fixed compilation warnings...

+++ BSR/Sources/BSR_Renderer.cpp
@@ -1082,7 +1082,7 @@ const bool BSR::Renderer::LoadModel(const char* _FileData, const size_t _FileSiz
 				{
 				case 0:
 				{
-					_FaceVertex.Position = strtoull(_Tokens[_Index], nullptr, 10) - 1;
+					_FaceVertex.Position = (size_t)(strtoull(_Tokens[_Index], nullptr, 10) - 1);
 
 					if (_FaceVertex.Position >= _Positions.GetSize())
 					{
@@ -1127,7 +1127,7 @@ const bool BSR::Renderer::LoadModel(const char* _FileData, const size_t _FileSiz
 						}
 					}
 
-					_FaceVertex.Position = strtoull(_FaceVertexTokens[0], nullptr, 10) - 1;
+					_FaceVertex.Position = (size_t)(strtoull(_FaceVertexTokens[0], nullptr, 10) - 1);
 
 					if (_FaceVertex.Position >= _Positions.GetSize())
 					{
@@ -1138,7 +1138,7 @@ const bool BSR::Renderer::LoadModel(const char* _FileData, const size_t _FileSiz
 						return false;
 					}
 
-					_FaceVertex.TextureCoords = strtoull(_FaceVertexTokens[1], nullptr, 10) - 1;
+					_FaceVertex.TextureCoords = (size_t)(strtoull(_FaceVertexTokens[1], nullptr, 10) - 1);
 
 					if (_FaceVertex.TextureCoords >= _TextureCoords.GetSize())
 					{
@@ -1185,7 +1185,7 @@ const bool BSR::Renderer::LoadModel(const char* _FileData, const size_t _FileSiz
 
 					if (_FaceVertexTokens.GetSize() == 2)
 					{
-						_FaceVertex.Position = strtoull(_FaceVertexTokens[0], nullptr, 10) - 1;
+						_FaceVertex.Position = (size_t)(strtoull(_FaceVertexTokens[0], nullptr, 10) - 1);
 
 						if (_FaceVertex.Position >= _Positions.GetSize())
 						{
@@ -1196,7 +1196,7 @@ const bool BSR::Renderer::LoadModel(const char* _FileData, const size_t _FileSiz
 							return false;
 						}
 
-						_FaceVertex.Normal = strtoull(_FaceVertexTokens[1], nullptr, 10) - 1;
+						_FaceVertex.Normal = (size_t)(strtoull(_FaceVertexTokens[1], nullptr, 10) - 1);
 
 						if (_FaceVertex.Normal >= _Normals.GetSize())
 						{
@@ -1209,7 +1209,7 @@ const bool BSR::Renderer::LoadModel(const char* _FileData, const size_t _FileSiz
 					}
 					else
 					{
-						_FaceVertex.Position = strtoull(_FaceVertexTokens[0], nullptr, 10) - 1;
+						_FaceVertex.Position = (size_t)(strtoull(_FaceVertexTokens[0], nullptr, 10) - 1);
 
 						if (_FaceVertex.Position >= _Positions.GetSize())
 						{
@@ -1220,7 +1220,7 @@ const bool BSR::Renderer::LoadModel(const char* _FileData, const size_t _FileSiz
 							return false;
 						}
 
-						_FaceVertex.TextureCoords = strtoull(_FaceVertexTokens[1], nullptr, 10) - 1;
+						_FaceVertex.TextureCoords = (size_t)(strtoull(_FaceVertexTokens[1], nullptr, 10) - 1);
 
 						if (_FaceVertex.TextureCoords >= _TextureCoords.GetSize())
 						{
@@ -1231,7 +1231,7 @@ const bool BSR::Renderer::LoadModel(const char* _FileData, const size_t _FileSiz
 							return false;
 						}
 
-						_FaceVertex.Normal = strtoull(_FaceVertexTokens[2], nullptr, 10) - 1;
+						_FaceVertex.Normal = (size_t)(strtoull(_FaceVertexTokens[2], nullptr, 10) - 1);
 
 						if (_FaceVertex.Normal >= _Normals.GetSize())
 						{
